% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milk.R
\docType{data}
\name{milk}
\alias{milk}
\title{Cleaned milk dataset}
\format{An object of class \code{data.frame} with 130 rows and 19 columns.}
\usage{
milk
}
\description{
Cleaned milk dataset
}
\keyword{datasets}
